requirements.txt
openai>=1.0.0
markitdown>=0.1.0
pyyaml>=6.0
httpx>=0.27.0
kafka-python>=2.0.0
pynetbox>=7.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
urllib3>=2.0.0
